refactor: Update product request count logic & text and move `RentalPeriodModel` to rental domain model (#163)
- Modify `GetProductDetailResultUseCase` to count requests that are not `CANCELED`, `REJECTED`, or `RETURNED` for the product owner.
- Change the string resource `screen_product_btn_request` from "요청 %d건" to "진행 중 %d건".
- Move `RentalPeriodModel` from `com.example.rentit.common.uimodel` to `com.example.rentit.domain.rental.model` and update imports accordingly.
