+++ app/src/main/java/com/example/rentit/domain/rental/model/RentalHistoryModel.kt
@@ -1,7 +1,6 @@
 package com.example.rentit.domain.rental.model
 
 import com.example.rentit.common.enums.RentalStatus
-import com.example.rentit.common.uimodel.RentalPeriodModel
 import java.time.LocalDateTime
 
 data class RentalHistoryModel(
